Mv menu rendering to new file
